moved turn to note command
