+++ src/main/java/frc/robot/VisionCommands/TurnToNote.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.VisionCommands;
 
 import java.util.function.DoubleSupplier;
 
@@ -10,6 +10,8 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.Constants;
+import frc.robot.SwerveConstants;
 import frc.robot.subsystems.NoteVision;
 import frc.robot.subsystems.Swerve;
